+++ src/main/java/DAJ2EE/demo/repository/UserRepository.java
@@ -4,10 +4,11 @@ import DAJ2EE.demo.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByUsername(String username);
+
     Optional<User> findByEmail(String email);
-    
+
     // Tìm kiếm bằng username (SĐT) HOẶC email để hỗ trợ đăng nhập linh hoạt
     Optional<User> findByUsernameOrEmail(String username, String email);
